Fixed [un]signed mismatch warnings. gc_root correctly destroyed on gc_destroy. Linux build fixed. Collision sample fixed. Removed (hopefully) last "new".

+++ src/mutiny/Resources.h
@@ -32,7 +32,7 @@ public:
     std::stringstream ss;
     ss << path << "_" << typeid(T).name();
 
-    for(int i = 0; i < Application::context->paths.size(); i++)
+    for(size_t i = 0; i < Application::context->paths.size(); i++)
     {
       if(ss.str() == Application::context->paths.at(i))
       {
